+++ install/sql/19-1.11.0-fix_forschung_notnull.sql
@@ -1,0 +1,7 @@
+-- MIT Licence
+-- Copyright (c) 2023 Simon Frankenberger
+-- Please see LICENCE.md for complete licence text.
+
+alter table mitglieder
+    modify Forschung13 smallint(2) default 0 not null,
+    modify Forschung14 smallint(2) default 0 not null;