+++ tests/src/test/resources/test-reset-player.php
@@ -52,6 +52,9 @@ switch ($testClass) {
         if ($testMethod === 'testInterestPlusWithCronAndBuilding') {
             Database::getInstance()->updateTableEntry(Database::TABLE_USERS, $id, array('Bank' => 99900, 'Gebaeude9' => 1));
         }
+        if ($testMethod === 'testResetAfterDispoLimit') {
+            Database::getInstance()->updateTableEntry(Database::TABLE_USERS, $id, array('Bank' => -109000));
+        }
         break;
 
     case 'BuildingTests':