+++ CHANGELOG.md
@@ -7,7 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-- no changes yet
+### Fixed
+
+- fix some minor inconsistencies for `Forschung`-columns in `mitglieder` database table
 
 ## [1.11.0] - 2023-01-28